# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget API",
        "description": "Documentación para Budget API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "admin@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/incomes": {
            "get": {
                "summary": "List all incomes",
                "operationId": "21d093ba630d134928aef66f298292d8",
                "responses": {
                    "200": {
                        "description": "A list of incomes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Store a new income",
                "operationId": "6a22317d9263a9d5bdc85795641c0268",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "date_ingreso",
                                    "type",
                                    "is_fixed"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "date_ingreso": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "is_fixed": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Income created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/incomes/{id}": {
            "get": {
                "summary": "Show a specific income",
                "operationId": "d2163801c56eb9e2a4ec97e300e0faba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Income details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a specific income",
                "operationId": "d320315cb025b9ceea6cf0df048ae7b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "date_ingreso",
                                    "type",
                                    "is_fixed"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "date_ingreso": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "is_fixed": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Income updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a specific income",
                "operationId": "62f7467011bc1a1b47cddedec4233011",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Income deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/reports/income-vs-egress": {
            "get": {
                "summary": "Get income vs egress report",
                "operationId": "271086697a709820710741ea518d579e",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Month to filter the report",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Year to filter the report",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Income vs Egress report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_income": {
                                            "type": "number"
                                        },
                                        "total_egress": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}